requirements.txt
asgiref==3.8.1
Django==4.2.17
psycopg2-binary==2.9.9
django-grappelli==4.0.2
django-user-agents==0.4.0
djangorestframework==3.16.0
sqlparse==0.5.3
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
user-agents==2.2.0
